backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
pydantic[email]==2.5.3
